Moved server's flight system to shared. I don't know why it isn't in the first place.
